refactor(ui): lift create-shelf dialog into ShelfDraftProvider
Mirrors UserSettingsDialogProvider: moves ShelfCreatorDialog + its mutation
into a dedicated context provider, so the sidebar button and the upcoming
command palette can both trigger it via useShelfDraftDialog().open().
